先拆分功能
