+++ lib/basic/logggg.go
@@ -1,0 +1,70 @@
+package basic
+
+import (
+	"fmt"
+	"os"
+	"path"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sirupsen/logrus"
+)
+
+func Logger() *logrus.Logger {
+	now := time.Now()
+	logFilePath := ""
+	if dir, err := os.Getwd(); err == nil {
+		logFilePath = dir + "/logs/"
+	}
+	if err := os.MkdirAll(logFilePath, 0777); err != nil {
+		fmt.Println(err.Error())
+	}
+	logFileName := now.Format("2006-01-02") + ".log"
+	//log文件
+	fileName := path.Join(logFilePath, logFileName)
+	if _, err := os.Stat(fileName); err != nil {
+		if _, err := os.Create(fileName); err != nil {
+			fmt.Println(err.Error())
+		}
+	}
+	//寫入文件
+	src, err := os.OpenFile(fileName, os.O_APPEND|os.O_WRONLY, os.ModeAppend)
+	if err != nil {
+		fmt.Println("err", err)
+	}
+
+	logger := logrus.New()
+	logger.Out = src
+	//設log等級
+	logger.SetLevel(logrus.DebugLevel)
+	//log格式
+	logger.SetFormatter(&logrus.TextFormatter{
+		TimestampFormat: "2006-01-02 15:04:05",
+	})
+	return logger
+}
+
+func LoggerToFile() gin.HandlerFunc {
+	logger := Logger()
+	return func(c *gin.Context) {
+		startTime := time.Now()
+		// 處理請求
+		c.Next()
+		endTime := time.Now()
+		// 執行時間
+		latencyTime := endTime.Sub(startTime)
+		// 請求方式
+		reqMethod := c.Request.Method
+		reqUri := c.Request.RequestURI
+		statusCode := c.Writer.Status()
+		clientIP := c.ClientIP()
+		//log格式
+		logger.Infof("| %3d | %13v | %15s | %s | %s |",
+			statusCode,
+			latencyTime,
+			clientIP,
+			reqMethod,
+			reqUri,
+		)
+	}
+}
